+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 5.0.0-beta7
+
+* The method Register on the interface Registry has been removed
+* SystemClock and ManualClock has been removed, as this is no longer need after 5.0.0-beta2
+* Logging class added
+
 ### 5.0.0-beta6
 
 * Fix bug in timer which could make the average zero in cases where the submonitors were read before the average